chore: getDemoById
