+++ examples/normal/.dumi/pages/loader-test.tsx
@@ -1,13 +1,13 @@
 import { useRouteData } from 'dumi';
-import * as demo from '../../docs/index.md?type=demo';
-import * as demoIndex from '../../docs/index.md?type=demo-index';
-import * as frontmatter from '../../docs/index.md?type=frontmatter';
-import * as text from '../../docs/index.md?type=text';
+// import * as demo from '../../docs/index.md?type=demo';
+// import * as demoIndex from '../../docs/index.md?type=demo-index';
+// import * as frontmatter from '../../docs/index.md?type=frontmatter';
+// import * as text from '../../docs/index.md?type=text';
 
-console.log('frontmatter', frontmatter);
-console.log('demo', demo);
-console.log('demoIndex', demoIndex);
-console.log('text', text);
+// console.log('frontmatter', frontmatter);
+// console.log('demo', demo);
+// console.log('demoIndex', demoIndex);
+// console.log('text', text);
 
 // Customize Page for dumi test
 export default () => {
