+++ examples/normal/.dumi/theme/layouts/GlobalLayout.tsx
@@ -1,14 +1,18 @@
-import {
-  useOutlet,
-  // useRouteData,
-  useRouteMeta,
-} from 'dumi';
+import { useOutlet } from 'dumi';
 import React from 'react';
-import { lazyMeta } from '../../tmp/dumi/meta/lazy';
+// import { lazyMeta } from '../../tmp/dumi/meta/lazy';
+import { getDemoById } from '../../tmp/dumi/meta/demos';
 
 export interface GlobalLayoutProps {
   children?: React.ReactNode;
 }
+
+console.log(
+  'getDemoById:',
+  getDemoById('foo-demo-0').then((ret) => {
+    console.log('Demo:', ret);
+  }),
+);
 
 export default function GlobalLayout({ children }: GlobalLayoutProps) {
   // console.log('site:', useSiteData());
@@ -20,13 +24,13 @@ export default function GlobalLayout({ children }: GlobalLayoutProps) {
   // console.log('useAppData:', useAppData());
   // console.log('useLocation:', useLocation());
   // console.log('useRouteData:', useRouteData());
-  const { id } = useRouteMeta();
+  // const { id } = useRouteMeta();
 
-  React.useEffect(() => {
-    lazyMeta(id).then((data) => {
-      console.log('Lazy Data:', data);
-    });
-  }, []);
+  // React.useEffect(() => {
+  //   lazyMeta(id).then((data) => {
+  //     console.log('Lazy Data:', data);
+  //   });
+  // }, []);
 
   const outlet = useOutlet();
